Refactor DB layer
- switch from sqlc to sqlx
- use repository pattern

+++ handlers/responses/crossonic.go
@@ -6,9 +6,9 @@ type ListenBrainzConfig struct {
 
 type Recap struct {
 	TotalDurationMS int64 `xml:"totalDurationMs" json:"totalDurationMs"`
-	SongCount       int64 `xml:"songCount" json:"songCount"`
-	AlbumCount      int64 `xml:"albumCount" json:"albumCount"`
-	ArtistCount     int64 `xml:"artistCount" json:"artistCount"`
+	SongCount       int   `xml:"songCount" json:"songCount"`
+	AlbumCount      int   `xml:"albumCount" json:"albumCount"`
+	ArtistCount     int   `xml:"artistCount" json:"artistCount"`
 }
 
 type TopSongsRecap struct {
